# hd: optimize ReverseBytes + add tests