@odeke-em 2017-07-27
/cc @ethanfrey @jaekwon for review please